#96 refactor-feature/todolist: 삭제 메서드 리팩토링 - @AuthenticationPrincipal을 추가하고CustomOAuth2User에서 memberName을 서비스로 넘김 - 맴버가 존재하는지, 해당 맴버가 해당 jd에 권한이 있는지 확인하는 로직 작성 - 삭제시 아무런 응답이 없는게 좋을 것 같은데 응답시 "" 으로 응답하여 null을 보내지 않음

+++ src/main/java/com/zb/jogakjogak/jobDescription/service/ToDoListService.java
@@ -102,16 +102,18 @@ public class ToDoListService {
      *
      * @param jdId       ToDoList가 속한 JD의 ID
      * @param toDoListId 조회할 ToDoList의 ID
-     * @return 삭제된 ToDoList의 응답 DTO
      */
-    public ToDoListDeleteResponseDto deleteToDoList(Long jdId, Long toDoListId) {
+    public void deleteToDoList(Long jdId, Long toDoListId, String memberName) {
 
+        Member member = memberRepository.findByUserName(memberName)
+                .orElseThrow(()-> new AuthException(MemberErrorCode.NOT_FOUND_MEMBER));
+        JD jd = findJdById(jdId);
+        if(!Objects.equals(jd.getMember().getId(),member.getId())){
+            throw new JDException(JDErrorCode.UNAUTHORIZED_ACCESS);
+        }
         ToDoList toDoList = findToDoListAndValidateOwnership(jdId, toDoListId);
 
         toDoListRepository.delete(toDoList);
-        return ToDoListDeleteResponseDto.builder()
-                .checklist_id(toDoList.getId())
-                .build();
     }
 
     /**
